refactor to separate file, simplify naming
