+++ apps/ensindexer/src/api/index.ts
@@ -2,39 +2,33 @@ import packageJson from "@/../package.json";
 
 import { db, publicClients } from "ponder:api";
 import schema from "ponder:schema";
-import {
-  createEnsRainbowVersionFetcher,
-  createFirstBlockToIndexByChainIdFetcher,
-  createPrometheusMetricsFetcher,
-  ensAdminUrl,
-  ensNodePublicUrl,
-  getEnsDeploymentChain,
-  getEnsDeploymentChainId,
-  ponderDatabaseSchema,
-  ponderPort,
-  requestedPluginNames,
-} from "@/lib/ponder-helpers";
-import {
-  PrometheusMetrics,
-  ponderMetadata,
-  queryPonderMeta,
-  queryPonderStatus,
-} from "@ensnode/ponder-metadata";
-import {
-  type Block,
-  buildGraphQLSchema,
-  graphql as subgraphGraphQL,
-} from "@ensnode/ponder-subgraph";
 import { Hono, MiddlewareHandler } from "hono";
 import { cors } from "hono/cors";
 import { client, graphql as ponderGraphQL } from "ponder";
-import type { PublicClient } from "viem";
+
+import {
+  ensAdminUrl,
+  ensNodePublicUrl,
+  getEnsDeploymentChain,
+  ponderDatabaseSchema,
+  requestedPluginNames,
+} from "@/lib/ponder-helpers";
+import {
+  fetchEnsRainbowVersion,
+  fetchFirstBlockToIndexByChainId,
+  fetchPrometheusMetrics,
+  makePonderMetdataProvider,
+} from "@/lib/ponder-metadata-provider";
+import { ponderMetadata } from "@ensnode/ponder-metadata";
+import {
+  buildGraphQLSchema as buildSubgraphGraphQLSchema,
+  graphql as subgraphGraphQL,
+} from "@ensnode/ponder-subgraph";
 
 const app = new Hono();
 
 const ensNodeVersionResponseHeader: MiddlewareHandler = async (ctx, next) => {
   ctx.header("x-ensnode-version", packageJson.version);
-
   return next();
 };
 
@@ -66,82 +60,6 @@ app.use("/", async (ctx) => {
     throw new Error(`Cannot redirect to ENSAdmin: ${errorMessage}`);
   }
 });
-
-// setup block indexing status fetching
-const fetchFirstBlockToIndexByChainId = createFirstBlockToIndexByChainIdFetcher(
-  import("../../ponder.config").then((m) => m.default),
-);
-
-// setup prometheus metrics fetching
-const fetchPrometheusMetrics = createPrometheusMetricsFetcher(ponderPort());
-
-// setup ENSRainbow version fetching
-const fetchEnsRainbowVersion = createEnsRainbowVersionFetcher();
-
-// setup the data provider for the ponder subgraph middleware
-const createPonderSubgraphDataProvider = () => {
-  // get the chain ID for the ENS deployment
-  const ensDeploymentChainId = getEnsDeploymentChainId();
-
-  /**
-   * Get the public client for the ENS deployment chain
-   * @returns the public client for the ENS deployment chain
-   * @throws an error if the public client is not found
-   */
-  function getEnsDeploymentPublicClient(): PublicClient {
-    // get the public client for the ENS deployment chain
-    const publicClient = publicClients[ensDeploymentChainId];
-
-    if (!publicClient) {
-      throw new Error(`Could not find public client for chain ID: ${ensDeploymentChainId}`);
-    }
-
-    return publicClient;
-  }
-
-  /**
-   * Get the last block indexed by Ponder.
-   *
-   * @returns the block info fetched from the public client
-   */
-  const getLastIndexedBlock = async (): Promise<Block> => {
-    const ponderStatus = await queryPonderStatus(ponderDatabaseSchema(), db);
-    const chainStatus = ponderStatus.find(
-      (status) => status.network_name === ensDeploymentChainId.toString(),
-    );
-
-    if (!chainStatus || !chainStatus.block_number) {
-      throw new Error(
-        `Could not find latest indexed block number for chain ID: ${ensDeploymentChainId}`,
-      );
-    }
-
-    return getEnsDeploymentPublicClient().getBlock({
-      blockNumber: BigInt(chainStatus.block_number),
-    });
-  };
-
-  /**
-   * Get the Ponder build ID
-   * @returns The Ponder build ID
-   */
-  const getPonderBuildId = async (): Promise<string> => {
-    const meta = await queryPonderMeta(ponderDatabaseSchema(), db);
-
-    return meta.build_id;
-  };
-
-  /**
-   * Check if there are any indexing errors logged in the prometheus metrics
-   * @returns true if there are no indexing errors, false otherwise
-   */
-  const hasIndexingErrors = async () => {
-    const metrics = PrometheusMetrics.parse(await fetchPrometheusMetrics());
-    return metrics.getValue("ponder_indexing_has_error") === 1;
-  };
-
-  return { getLastIndexedBlock, getPonderBuildId, hasIndexingErrors };
-};
 
 // use ENSNode middleware at /metadata
 app.get(
@@ -177,12 +95,10 @@ app.use(
   "/subgraph",
   subgraphGraphQL({
     db,
-    graphqlSchema: buildGraphQLSchema({
+    graphqlSchema: buildSubgraphGraphQLSchema({
       schema,
-      metaConfig: {
-        version: packageJson.version,
-        schema: ponderDatabaseSchema(),
-      },
+      // provide the schema with ponder's internal metadata to power _meta
+      metadataProvider: makePonderMetdataProvider(),
       // describes the polymorphic (interface) relationships in the schema
       polymorphicConfig: {
         types: {
@@ -217,7 +133,6 @@ app.use(
           "Resolver.events": "ResolverEvent",
         },
       },
-      dataProvider: createPonderSubgraphDataProvider(),
     }),
   }),
 );
